refactor: eventlisteners separated by their names and roles
